search algos fix 0

+++ 0x1E-search_algorithms/0-linear.c
@@ -8,7 +8,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (!array)
 		return (-1);
@@ -16,7 +16,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		if (array[i] == value)
 			return (i);
-		printf("Value checked array[%d] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
 	}
 
 	return (-1);
